refactor(sandbox): lift the player.style harness out of the instaplay page
The comparison shell, both panels and the router were written inside the
instaplay template. Thirteen more themes are coming, so they move to
`app/shared/player-style/` and a theme becomes markup, a stylesheet and two
small entries.

`themes.ts` is the registry: every theme's preset and the published version its
reference frame pins. Pinned rather than a range, since bumping one changes
what the comparison compares against.

Each panel entry now imports `@app/styles.css` itself instead of relying on a
caller, and `elements.ts` registers the presets alongside the UI elements. Both
were latent in the single-page version, where one entry happened to import
everything.
