+++ apps/sandbox/app/shared/player-style/panel-videojs.ts
@@ -1,0 +1,43 @@
+import '@app/styles.css';
+import { defineTemplateSkin } from '@app/shared/html/registry-skins';
+import { createHtmlSandbox, html } from '@app/shared/html/sandbox';
+
+import { type PlayerStyleTheme, portedTagName } from './config';
+
+/** The frame around a player, matching what the stock sandbox templates use. */
+export function frameClasses({ player }: PlayerStyleTheme): string {
+  return player === 'audio' ? 'mx-auto w-full max-w-xl' : 'mx-auto aspect-video w-full max-w-4xl';
+}
+
+/**
+ * Mount the ported theme.
+ *
+ * The theme is markup plus a stylesheet, stamped into an element the same way an ejected registry skin is: the authored
+ * markup keeps a bare `<slot>` where the media goes and the poster's own `<img>` as its target. The page names its own
+ * tag rather than the shell's `skinTag`, since the shell's picker only knows the packaged skins.
+ */
+export function mountPortedPanel(theme: PlayerStyleTheme, markup: string): void {
+  const skinTag = defineTemplateSkin(portedTagName(theme), {
+    markup,
+    media: (container) => container.querySelector('slot:not([name])'),
+    poster: (container) => container.querySelector('media-poster img'),
+  });
+
+  const playerTag = theme.player === 'audio' ? 'audio-player' : 'video-player';
+  const mediaTag = theme.player === 'audio' ? 'audio' : 'video';
+
+  createHtmlSandbox({
+    player: theme.player,
+    render: ({ src, attrs, chapters, storyboard, poster }) => html`
+      <${playerTag}>
+        <${skinTag} class="${frameClasses(theme)}">
+          <${mediaTag}${src} ${attrs} playsinline crossorigin>
+            ${chapters}
+            ${storyboard}
+          </${mediaTag}>
+          ${poster ? html`<img slot="poster" src="${poster}" alt="Video poster" crossorigin />` : ''}
+        </${skinTag}>
+      </${playerTag}>
+    `,
+  });
+}
